fix(anyharness): satisfy PR2 repo shape checks
