+++ anyharness/crates/anyharness-lib/src/domains/workspaces/options.rs
@@ -539,216 +539,9 @@ fn repository_name(root: &RepoRootRecord) -> String {
         .unwrap_or_else(|| "repository".to_string())
 }
 
-fn validate_workspace_creation(
-    options: &WorkspaceCreationOptions,
-    input: &CreateWorkspaceFromOptionsInput,
-) -> Result<ValidatedWorkspaceCreation, WorkspaceOptionsError> {
-    let repository_id = input.repository_id.trim();
-    let repository = options
-        .repository(repository_id)
-        .ok_or_else(|| WorkspaceOptionsError::RepositoryNotFound(repository_id.to_string()))?;
-    if !repository.availability.is_present() {
-        return Err(WorkspaceOptionsError::RepositoryUnavailable {
-            repository_id: repository_id.to_string(),
-            availability: repository.availability.clone(),
-        });
-    }
-
-    let creation_mode = WorkspaceCreationMode::parse(&input.creation_mode)?;
-    let branch = match creation_mode {
-        WorkspaceCreationMode::Worktree => {
-            let branch = input
-                .branch
-                .as_deref()
-                .map(str::trim)
-                .filter(|branch| !branch.is_empty())
-                .ok_or(WorkspaceOptionsError::BranchRequired)?;
-            validate_new_branch(branch).map_err(WorkspaceOptionsError::InvalidBranch)?;
-            Some(branch.to_string())
-        }
-        WorkspaceCreationMode::Local => {
-            if input.branch.is_some() {
-                return Err(WorkspaceOptionsError::BranchNotAllowed);
-            }
-            None
-        }
-    };
-    let display_name = normalize_workspace_display_name(input.display_name.as_deref())
-        .map_err(map_display_name_error)?;
-
-    Ok(ValidatedWorkspaceCreation {
-        repository_id: repository_id.to_string(),
-        creation_mode,
-        branch,
-        display_name,
-    })
-}
-
-fn validate_new_branch(branch: &str) -> Result<(), String> {
-    if branch.starts_with('-') {
-        return Err("branch name must not begin with '-'".into());
-    }
-    if branch == "HEAD" {
-        return Err("branch name must not be 'HEAD'".into());
-    }
-    if branch.starts_with('/') || branch.ends_with('/') || branch.contains("//") {
-        return Err("branch name has an invalid '/' component".into());
-    }
-    if branch.ends_with('.') || branch.contains("..") {
-        return Err("branch name must not contain '..' or end with '.'".into());
-    }
-    if branch.contains("@{") || branch == "@" {
-        return Err("branch name contains a forbidden ref sequence".into());
-    }
-    for component in branch.split('/') {
-        if component.starts_with('.') || component.ends_with(".lock") {
-            return Err("branch name has an invalid component".into());
-        }
-    }
-    if branch.chars().any(|character| {
-        character.is_ascii_control()
-            || matches!(
-                character,
-                ' ' | '~' | '^' | ':' | '?' | '*' | '[' | '\\' | '\u{7f}'
-            )
-    }) {
-        return Err("branch name contains a forbidden character".into());
-    }
-    Ok(())
-}
-
-fn map_display_name_error(error: SetWorkspaceDisplayNameError) -> WorkspaceOptionsError {
-    match error {
-        SetWorkspaceDisplayNameError::TooLong(limit) => {
-            WorkspaceOptionsError::DisplayNameTooLong(limit)
-        }
-        SetWorkspaceDisplayNameError::NotFound(workspace_id) => {
-            WorkspaceOptionsError::WorkspaceNotFound(workspace_id)
-        }
-        SetWorkspaceDisplayNameError::Unexpected(error) => WorkspaceOptionsError::Create(error),
-    }
-}
+mod validation;
+use validation::{map_display_name_error, validate_workspace_creation};
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn options(availability: WorkspaceRepositoryAvailability) -> WorkspaceCreationOptions {
-        WorkspaceCreationOptions {
-            repositories: vec![WorkspaceRepositoryOption {
-                repository_id: "repo-1".into(),
-                name: "Repo".into(),
-                path: "/repo".into(),
-                default_branch: Some("main".into()),
-                current_branch: Some("main".into()),
-                branches: Vec::new(),
-                executable: availability.is_present(),
-                unavailable_reason: availability.public_reason(),
-                availability,
-            }],
-            creation_modes: Vec::new(),
-        }
-    }
-
-    fn input(mode: &str, branch: Option<&str>) -> CreateWorkspaceFromOptionsInput {
-        CreateWorkspaceFromOptionsInput {
-            repository_id: "repo-1".into(),
-            creation_mode: mode.into(),
-            branch: branch.map(str::to_string),
-            display_name: None,
-            origin: OriginContext::system_local_runtime(),
-            creator_context: WorkspaceCreatorContext::Agent {
-                source_session_id: "session-1".into(),
-                source_session_workspace_id: Some("workspace-1".into()),
-                session_link_id: None,
-                source_workspace_id: Some("workspace-1".into()),
-                label: None,
-            },
-        }
-    }
-
-    #[test]
-    fn shared_validator_accepts_the_listed_modes_and_enforces_branch_shape() {
-        let available = options(WorkspaceRepositoryAvailability::Present);
-        let worktree =
-            validate_workspace_creation(&available, &input("worktree", Some("feature/agent-ops")))
-                .expect("listed worktree choice");
-        assert_eq!(worktree.creation_mode, WorkspaceCreationMode::Worktree);
-        assert_eq!(worktree.branch.as_deref(), Some("feature/agent-ops"));
-
-        assert!(matches!(
-            validate_workspace_creation(&available, &input("worktree", None)),
-            Err(WorkspaceOptionsError::BranchRequired)
-        ));
-        assert!(matches!(
-            validate_workspace_creation(&available, &input("local", Some("main"))),
-            Err(WorkspaceOptionsError::BranchNotAllowed)
-        ));
-        assert!(validate_workspace_creation(&available, &input("local", None)).is_ok());
-    }
-
-    #[test]
-    fn stale_missing_and_unreadable_repositories_are_typed_rejections() {
-        for availability in [
-            WorkspaceRepositoryAvailability::Missing,
-            WorkspaceRepositoryAvailability::Unreadable {
-                diagnostic: "permission denied at /private/runtime/repository".into(),
-            },
-        ] {
-            let error =
-                validate_workspace_creation(&options(availability.clone()), &input("local", None))
-                    .expect_err("unavailable option must be rejected");
-            assert!(matches!(
-                error,
-                WorkspaceOptionsError::RepositoryUnavailable {
-                    availability: observed,
-                    ..
-                } if observed == availability
-            ));
-        }
-    }
-
-    #[test]
-    fn unreadable_repository_serializes_only_a_stable_state_and_safe_message() {
-        let raw_diagnostic =
-            "fatal: could not read /private/customer/repository/.git: permission denied";
-        let availability = WorkspaceRepositoryAvailability::Unreadable {
-            diagnostic: raw_diagnostic.to_string(),
-        };
-        let projected = options(availability.clone());
-        let json = serde_json::to_value(&projected).expect("serialize workspace options");
-        assert_eq!(
-            json["repositories"][0]["availability"]["state"],
-            "unreadable"
-        );
-        assert_eq!(
-            json["repositories"][0]["unavailableReason"],
-            "The repository checkout could not be read."
-        );
-        let serialized = json.to_string();
-        assert!(!serialized.contains(raw_diagnostic));
-        assert!(!serialized.contains("/private/customer"));
-
-        let error = WorkspaceOptionsError::RepositoryUnavailable {
-            repository_id: "repo-1".to_string(),
-            availability,
-        };
-        assert_eq!(error.code(), "WORKSPACE_REPOSITORY_UNAVAILABLE");
-        let public = error.public_message();
-        assert_eq!(public, "The repository checkout could not be read.");
-        assert!(!public.contains("permission denied"));
-        assert!(!public.contains("/private/customer"));
-    }
-
-    #[test]
-    fn existing_branch_metadata_is_explicitly_not_a_creation_token() {
-        let metadata = WorkspaceBranchMetadata::from(GitBranch {
-            name: "main".into(),
-            is_remote: false,
-            is_head: true,
-            is_default: true,
-            upstream: Some("origin/main".into()),
-        });
-        assert!(!metadata.selectable_for_creation);
-    }
-}
+#[path = "options/tests.rs"]
+mod tests;
